requirements.txt
# SenseType 依赖
# Python 3.11.14 (micromamba env: winvoice)
#
# 部署步骤:
#   micromamba create -n winvoice python=3.11
#   micromamba activate winvoice
#   pip install torch==2.5.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu121
#   pip install -r requirements.txt

# --- ASR ---
funasr==1.3.1
modelscope==1.34.0
sentencepiece==0.2.1

# --- 录音 ---
sounddevice==0.5.5

# --- 热键 ---
keyboard==0.13.5

# --- 录音保存 ---
soundfile==0.13.1

# --- 系统托盘 ---
pystray==0.19.5
pillow==11.0.0

# --- 基础依赖 ---
numpy==2.4.2
